+++ src/components/input/UserInput.tsx
@@ -105,33 +105,32 @@ export default function UserInput(props: UserConfigProps) {
         <Form.Row>
           <Form.Group as={Col} controlId="formHeaderSize">
             <HelpTooltip
-              text="For a LoRaWAN 1.0.x uplink and downlink, the header is at least 13 bytes: Type (1), DevAddr (4), FCtrl (1), FCnt (2), FPort (1) and MIC (4)."
-              placement="top"
-              children={<Form.Label>Header size</Form.Label>}
-            />
+              showIcon={true}
+              content="For a LoRaWAN 1.0.x uplink and downlink, the header is at least 13 bytes: Type (1), DevAddr (4), FCtrl (1), FCnt (2), FPort (1) and MIC (4)."
+            >
+              <Form.Label>Header size</Form.Label>
+            </HelpTooltip>
             <NumberInput {...headerSize} min={13} />
           </Form.Group>
 
           <Form.Group as={Col} controlId="formApplicationPayloadSize">
             <HelpTooltip
-              text="The application payload size. Might be empty for a simple ACK or if the packet only includes MAC commands."
-              placement="top"
-              children={<Form.Label>Payload size</Form.Label>}
-            />
+              showIcon={true}
+              content="The application payload size. Might be empty for a simple ACK or if the packet only includes MAC commands."
+            >
+              <Form.Label>Payload size</Form.Label>
+            </HelpTooltip>
             <NumberInput {...payloadSize} />
           </Form.Group>
 
           <Form.Group as={Col} controlId="formCodingRate">
             <HelpTooltip
-              text="The coding rate (CR) used for forward error correction (FEC)."
-              placement="top"
-              children={<Form.Label>Coding rate</Form.Label>}
-            />
-            <Form.Control
-              as="select"
-              {...withFormControl(codingRate)}
-              className="xxtext-primary xxborder-primary"
+              showIcon={true}
+              content="The coding rate (CR) used for forward error correction (FEC)."
             >
+              <Form.Label>Coding rate</Form.Label>
+            </HelpTooltip>
+            <Form.Control as="select" {...withFormControl(codingRate)}>
               <option value="4/5">4/5</option>
               <option value="4/6">4/6</option>
               <option value="4/7">4/7</option>