+++ src/components/result/Result.tsx
@@ -1,23 +1,44 @@
 import React from 'react';
-import OverlayTrigger from 'react-bootstrap/OverlayTrigger';
-import Tooltip from 'react-bootstrap/Tooltip';
 import {DataRate} from '../../AppConfig';
+import HelpTooltip from '../help/HelpTooltip';
 
 type ResultGridProps = {
   dr: DataRate;
+  size: number;
   airtime: number;
+  maxPhyPayloadSize: number;
   maxDwellTime?: number;
 };
 
-function fmt(n: number, digits: number) {
+export function fmt(n: number, digits: number) {
   return n.toLocaleString('en-US', {minimumFractionDigits: digits, maximumFractionDigits: digits});
+}
+
+/**
+ * Join non-empty fragments into a sentence, adding full stops when combining
+ * more than one fragment.
+ */
+export function withFullStops(...lines: (string | false | 0 | undefined)[]) {
+  return (
+    lines.reduce((acc: string, line) => {
+      if (!line) {
+        return acc;
+      }
+      return (
+        acc +
+        (acc && acc.slice(-1) !== '.' ? '. ' : acc ? ' ' : '') +
+        line +
+        (acc && line.slice(-1) !== '.' ? '.' : '')
+      );
+    }, '') || undefined
+  );
 }
 
 /**
  * A single result, showing the data rate (SF, bandwidth), time on air and
  * calculated limits.
  */
-export function Result({dr, airtime, maxDwellTime = 0}: ResultGridProps) {
+export function Result({dr, size, airtime, maxPhyPayloadSize, maxDwellTime = 0}: ResultGridProps) {
   // Minimum delay in seconds between two packet starts, for maximum 1% duty cycle
   const dcDelay = airtime / 0.01 / 1000;
 
@@ -25,17 +46,25 @@ export function Result({dr, airtime, maxDwellTime = 0}: ResultGridProps) {
   const fapMessages = 30000 / airtime;
   const fapDelay = (24 * 3600) / fapMessages;
 
-  const tooLong = maxDwellTime > 0 && airtime > maxDwellTime;
+  const dwellTimeExceeded = maxDwellTime > 0 && airtime > maxDwellTime;
+  const sizeExceeded = dr.maxMacPayloadSize && size > maxPhyPayloadSize;
 
   return (
     <>
-      <OverlayTrigger
-        placement={`top`}
-        overlay={
-          dr.notes ? <Tooltip id={`Result-datarate-${dr.name}-help`}>{dr.notes}</Tooltip> : <span />
-        }
+      <HelpTooltip
+        content={withFullStops(
+          dr.notes,
+          dr.maxMacPayloadSize &&
+            (sizeExceeded
+              ? `The full payload of ${size} bytes exceeds the maximum packet size of ${maxPhyPayloadSize} bytes`
+              : `Maximum total packet size ${maxPhyPayloadSize} bytes`)
+        )}
       >
-        <div className={`Result-datarate ${dr.notes ? 'Result-has-note' : ''}`}>
+        <div
+          className={`Result-datarate ${
+            sizeExceeded ? 'Result-has-warning' : dr.notes ? 'Result-has-note' : ''
+          }`}
+        >
           <div className="Result-dr">{dr.name}</div>
           <div>
             <span className="Result-sf">SF{dr.sf}</span>
@@ -45,35 +74,34 @@ export function Result({dr, airtime, maxDwellTime = 0}: ResultGridProps) {
               {dr.bw}
             </span>
           </div>
+          {sizeExceeded && <div className={'Result-datarate-warning'}>max size exceeded</div>}
         </div>
-      </OverlayTrigger>
+      </HelpTooltip>
 
-      <OverlayTrigger
-        placement={`top`}
-        overlay={
-          <Tooltip id="Result-dutycycle-help">
-            {fmt(airtime, 3)} milliseconds time on air
-            {tooLong && `. This exceeds the maximum dwell time of ${maxDwellTime} milliseconds.`}
-          </Tooltip>
-        }
+      <HelpTooltip
+        content={withFullStops(
+          `${fmt(airtime, 3)} milliseconds time on air`,
+          dwellTimeExceeded && `This exceeds the maximum dwell time of ${maxDwellTime} milliseconds`
+        )}
       >
-        <div className={`Result-airtime ${tooLong ? 'Result-has-warning' : ''}`}>
-          <span>
+        <div className={`Result-airtime ${dwellTimeExceeded ? 'Result-has-warning' : ''}`}>
+          <div>
             {fmt(airtime, 1)}
             <span className="Result-unit">ms</span>
-          </span>
-          {tooLong && <span className={'Result-dwelltime-warning'}>max dwell time exceeded</span>}
+          </div>
+          {dwellTimeExceeded && (
+            <div className={'Result-airtime-warning'}>max dwell time exceeded</div>
+          )}
         </div>
-      </OverlayTrigger>
+      </HelpTooltip>
 
-      <OverlayTrigger
-        placement={`top`}
-        overlay={
-          <Tooltip id="Result-dutycycle-help">
+      <HelpTooltip
+        content={
+          <>
             If regional legal or LoRaWAN maximum duty cycles apply, then a 1% maximum duty cycle
             needs a <em>minimum</em> of {fmt(dcDelay, 3)} seconds between <em>any</em> subsequent
             packet starts in the same subband.
-          </Tooltip>
+          </>
         }
       >
         <div className="Result-dutycycle">
@@ -90,12 +118,11 @@ export function Result({dr, airtime, maxDwellTime = 0}: ResultGridProps) {
             </span>
           </div>
         </div>
-      </OverlayTrigger>
+      </HelpTooltip>
 
-      <OverlayTrigger
-        placement={`top`}
-        overlay={
-          <Tooltip id="Result-dutycycle-help">
+      <HelpTooltip
+        content={
+          <>
             <p>
               The TTN Fair Access Policy allows for at most 30 seconds time on air per device, per
               24 hours. An airtime of {fmt(airtime, 3)}&nbsp;ms thus yields a maximum of{' '}
@@ -106,7 +133,7 @@ export function Result({dr, airtime, maxDwellTime = 0}: ResultGridProps) {
               {fmt(fapDelay, 1)} seconds between the starts of two uplinks, or a maximum of{' '}
               {fmt(fapMessages / 24, 1)}&nbsp;messages per hour.
             </p>
-          </Tooltip>
+          </>
         }
       >
         <div className="Result-fairaccess">
@@ -135,7 +162,7 @@ export function Result({dr, airtime, maxDwellTime = 0}: ResultGridProps) {
             </span>
           </div>
         </div>
-      </OverlayTrigger>
+      </HelpTooltip>
     </>
   );
 }