+++ src/setupTests.ts
@@ -16,6 +16,7 @@ export class Fixtures {
     name: 'DR6',
     sf: 7,
     bw: 250,
+    maxMacPayloadSize: 230,
     highlight: 'low',
   };
 
@@ -23,41 +24,48 @@ export class Fixtures {
     name: 'DR5',
     sf: 7,
     bw: 125,
+    maxMacPayloadSize: 230,
   };
 
   static eu868dr4: DataRate = {
     name: 'DR4',
     sf: 8,
     bw: 125,
+    maxMacPayloadSize: 230,
   };
 
   static eu868dr3: DataRate = {
     name: 'DR3',
     sf: 9,
     bw: 125,
+    maxMacPayloadSize: 123,
   };
 
   static eu868dr2: DataRate = {
     name: 'DR2',
     sf: 10,
     bw: 125,
+    maxMacPayloadSize: 59,
   };
 
   static eu868dr1: DataRate = {
     name: 'DR1',
     sf: 11,
     bw: 125,
+    maxMacPayloadSize: 59,
   };
 
   static eu868dr0: DataRate = {
     name: 'DR0',
     sf: 12,
     bw: 125,
+    maxMacPayloadSize: 59,
   };
 
   static eu868: Region = {
     name: 'eu868',
     label: 'EU868',
+    title: 'EU863-870 uplink and downlink',
     dataRates: [
       Fixtures.eu868dr6,
       Fixtures.eu868dr5,