+++ CHANGELOG.md
@@ -1,10 +1,14 @@
 ## [1.0.1](https://github.com/avbentem/airtime-calculator/compare/v1.0.0...master) (future)
 
+### New features
+
+- Validate against LoRaWAN-defined maximum payload size.
+
 ### Fixes
 
 - Document 1% duty cycle for the default channels of AS923.
 - Document usage of AS923 in Australia.
-- De-emphasize/shade all data rates that only apply to a single-frequency.
+- De-emphasize/shade all data rates that only apply to a single frequency.
 - Layout fixes.
 
 ## 1.0.0 (2020-07-21)